scraper/requirements.txt
scrapy==2.11.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
python-dotenv==1.0.0
Sastrawi==1.0.1
